// lib/typings/types.ts
  export type Variables = {
    taskId: string;
  };

  export type Query = {
    __typename?: "Query";

    SdmGoal?: (SdmGoal | null)[] | null;
  };

  export type SdmGoal = {
    __typename?: "SdmGoal";

    data?: string | null;

    repo?: Repo | null;
  };

  export type Repo = {
    __typename?: "SdmRepository";

    name?: string | null;

    owner?: string | null;

    providerId?: string | null;
  };
// ...
export namespace SonarScanCompleted {
  export type Variables = {};

  export type Subscription = {
    __typename?: "Subscription";

    SonarScan?: (SonarScan | null)[] | null;
  };

  export type SonarScan = {
    __typename?: "SonarScan";

    project: Project;

    push?: Push | null;

    qualityGate: QualityGate;

    status: string;

    taskId: string;
  };

  export type Project = {
    __typename?: "SonarProject";

    name?: string | null;

    key?: string | null;
  };

  export type Push = {
    __typename?: "Push";

    branch?: string | null;

    commits?: (Commits | null)[] | null;
  };

  export type Commits = {
    __typename?: "Commit";

    sha?: string | null;

    repo?: Repo | null;
  };

  export type Repo = {
    __typename?: "Repo";

    owner?: string | null;

    name?: string | null;

    channels?: (Channels | null)[] | null;
  };

  export type Channels = {
    __typename?: "ChatChannel";

    name?: string | null;
  };

  export type QualityGate = {
    __typename?: "SonarQualityGate";

    conditions?: (Conditions | null)[] | null;
  };

  export type Conditions = {
    __typename?: "SonarQualityGateCondition";

    errorThreshold?: string | null;

    metric?: string | null;

    onLeakPeriod?: boolean | null;

    operator?: string | null;

    status?: string | null;

    value?: string | null;
  };
}